# Conflicts:
#	library.properties
#	support/Release notes.txt
#	test/codeblocks/ETL.layout

+++ src/type_lookup.h
@@ -101,7 +101,7 @@ namespace etl
     template <int ID>
     struct type_from_id
     {
-      typedef 
+      typedef
             typename etl::conditional<ID == T1::ID, typename T1::type,
             typename etl::conditional<ID == T2::ID, typename T2::type,
             typename etl::conditional<ID == T3::ID, typename T3::type,
@@ -138,22 +138,22 @@ namespace etl
       enum
       {
         value =
-          (unsigned int) etl::is_same<T, typename T1::type>::value ? T1::ID :
-          (unsigned int) etl::is_same<T, typename T2::type>::value ? T2::ID :
-          (unsigned int) etl::is_same<T, typename T3::type>::value ? T3::ID :
-          (unsigned int) etl::is_same<T, typename T4::type>::value ? T4::ID :
-          (unsigned int) etl::is_same<T, typename T5::type>::value ? T5::ID :
-          (unsigned int) etl::is_same<T, typename T6::type>::value ? T6::ID :
-          (unsigned int) etl::is_same<T, typename T7::type>::value ? T7::ID :
-          (unsigned int) etl::is_same<T, typename T8::type>::value ? T8::ID :
-          (unsigned int) etl::is_same<T, typename T9::type>::value ? T9::ID :
-          (unsigned int) etl::is_same<T, typename T10::type>::value ? T10::ID :
-          (unsigned int) etl::is_same<T, typename T11::type>::value ? T11::ID :
-          (unsigned int) etl::is_same<T, typename T12::type>::value ? T12::ID :
-          (unsigned int) etl::is_same<T, typename T13::type>::value ? T13::ID :
-          (unsigned int) etl::is_same<T, typename T14::type>::value ? T14::ID :
-          (unsigned int) etl::is_same<T, typename T15::type>::value ? T15::ID :
-          (unsigned int) etl::is_same<T, typename T16::type>::value ? T16::ID :
+          (unsigned int) etl::is_same<T, typename T1::type>::value ? (unsigned int)T1::ID :
+          (unsigned int) etl::is_same<T, typename T2::type>::value ? (unsigned int)T2::ID :
+          (unsigned int) etl::is_same<T, typename T3::type>::value ? (unsigned int)T3::ID :
+          (unsigned int) etl::is_same<T, typename T4::type>::value ? (unsigned int)T4::ID :
+          (unsigned int) etl::is_same<T, typename T5::type>::value ? (unsigned int)T5::ID :
+          (unsigned int) etl::is_same<T, typename T6::type>::value ? (unsigned int)T6::ID :
+          (unsigned int) etl::is_same<T, typename T7::type>::value ? (unsigned int)T7::ID :
+          (unsigned int) etl::is_same<T, typename T8::type>::value ? (unsigned int)T8::ID :
+          (unsigned int) etl::is_same<T, typename T9::type>::value ? (unsigned int)T9::ID :
+          (unsigned int) etl::is_same<T, typename T10::type>::value ? (unsigned int)T10::ID :
+          (unsigned int) etl::is_same<T, typename T11::type>::value ? (unsigned int)T11::ID :
+          (unsigned int) etl::is_same<T, typename T12::type>::value ? (unsigned int)T12::ID :
+          (unsigned int) etl::is_same<T, typename T13::type>::value ? (unsigned int)T13::ID :
+          (unsigned int) etl::is_same<T, typename T14::type>::value ? (unsigned int)T14::ID :
+          (unsigned int) etl::is_same<T, typename T15::type>::value ? (unsigned int)T15::ID :
+          (unsigned int) etl::is_same<T, typename T16::type>::value ? (unsigned int)T16::ID :
           (unsigned int) UNKNOWN
       };
 
@@ -202,7 +202,7 @@ namespace etl
     template <typename T>
     struct type_from_type
     {
-      typedef 
+      typedef
             typename etl::conditional<etl::is_same<T, typename T1::type1>::value, typename T1::type2,
             typename etl::conditional<etl::is_same<T, typename T2::type1>::value, typename T2::type2,
             typename etl::conditional<etl::is_same<T, typename T3::type1>::value, typename T3::type2,
